# `<bitset>`: optimize streaming `operator >>` 