--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -608,42 +608,45 @@ basic_istream<_Elem, _Tr>& operator>>(basic_istream<_Elem, _Tr>& _Istr, bitset<_
     const _Elem _Elem0               = _Ctype_fac.widen('0');
     const _Elem _Elem1               = _Ctype_fac.widen('1');
     typename _Istr_t::iostate _State = _Istr_t::goodbit;
-    bool _Changed                    = false;
+    size_t _Count                    = 0;
     string _Str;
     const typename _Istr_t::sentry _Ok(_Istr);
 
     if (_Ok) { // valid stream, extract elements
         _TRY_IO_BEGIN
-        typename _Tr::int_type _Meta = _Istr.rdbuf()->sgetc();
-        for (size_t _Count = _Right.size(); 0 < _Count; _Meta = _Istr.rdbuf()->snextc(), (void) --_Count) {
-            // test _Meta
-            if (_Tr::eq_int_type(_Tr::eof(), _Meta)) { // end of file, quit
-                _State |= _Istr_t::eofbit;
-                break;
+        _Str._Resize_and_overwrite(_Bits, [&, _Elem0, _Elem1](char* _Buf, size_t _Max_count) {
+            typename _Tr::int_type _Meta = _Istr.rdbuf()->sgetc();
+            for (; _Count < _Max_count; _Meta = _Istr.rdbuf()->snextc(), (void) ++_Count) {
+                // test _Meta
+                if (_Tr::eq_int_type(_Tr::eof(), _Meta)) { // end of file, quit
+                    _State |= _Istr_t::eofbit;
+                    break;
+                }
+
+                const _Elem _Char = _Tr::to_char_type(_Meta);
+
+                if (!_Tr::eq(_Char, _Elem0) && !_Tr::eq(_Char, _Elem1)) {
+                    break; // invalid element
+                }
+
+                if (_Str.max_size() <= _Str.size()) { // no room in string, give up (unlikely)
+                    _State |= _Istr_t::failbit;
+                    break;
+                }
+
+                // valid, append '0' or '1'
+                _Buf[_Count] = '0' + _Tr::eq(_Char, _Elem1);
             }
 
-            const _Elem _Char = _Tr::to_char_type(_Meta);
-
-            if (!_Tr::eq(_Char, _Elem0) && !_Tr::eq(_Char, _Elem1)) {
-                break; // invalid element
-            }
-
-            if (_Str.max_size() <= _Str.size()) { // no room in string, give up (unlikely)
-                _State |= _Istr_t::failbit;
-                break;
-            }
-
-            // valid, append '0' or '1'
-            _Str.push_back('0' + _Tr::eq(_Char, _Elem1));
-            _Changed = true;
-        }
+            return _Count;
+        });
         _CATCH_IO_(_Istr_t, _Istr)
     }
 
     constexpr bool _Has_bits = _Bits > 0;
 
     if constexpr (_Has_bits) {
-        if (!_Changed) {
+        if (_Count != 0) {
             _State |= _Istr_t::failbit;
         }
     }