--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -40,6 +40,36 @@ __declspec(noalias) bool __stdcall __std_bitset_from_string_2(void* _Dest, const
 #endif // _USE_STD_VECTOR_ALGORITHMS
 
 _STD_BEGIN
+template <class _Ty, size_t _Size, bool _Dynamic = (_Size > _STD_BITSET_TO_STREAM_STACK_RESERVATION)>
+class _Buffer_for_bitset {
+public:
+    _Buffer_for_bitset()                                     = default;
+    _Buffer_for_bitset(const _Buffer_for_bitset&)            = delete;
+    _Buffer_for_bitset& operator=(const _Buffer_for_bitset&) = delete;
+
+    ~_Buffer_for_bitset() {
+        delete[] _Buf;
+    }
+
+    void _Allocate() {
+        _Buf = new _Ty[_Size];
+    }
+
+    _Ty* _Buf = nullptr;
+};
+
+template <class _Ty, size_t _Size>
+class _Buffer_for_bitset<_Ty, _Size, false> {
+public:
+    _Buffer_for_bitset() { /* not initializing _Buf intentionally */ }
+    _Buffer_for_bitset(const _Buffer_for_bitset&)            = delete;
+    _Buffer_for_bitset& operator=(const _Buffer_for_bitset&) = delete;
+
+    void _Allocate() {}
+
+    _Ty _Buf[_STD_BITSET_TO_STREAM_STACK_RESERVATION];
+};
+
 _EXPORT_STD template <size_t _Bits>
 class bitset { // store fixed-length sequence of Boolean elements
 private:
@@ -609,37 +639,30 @@ basic_istream<_Elem, _Tr>& operator>>(basic_istream<_Elem, _Tr>& _Istr, bitset<_
     const _Elem _Elem1               = _Ctype_fac.widen('1');
     typename _Istr_t::iostate _State = _Istr_t::goodbit;
     size_t _Count                    = 0;
-    string _Str;
+    _Buffer_for_bitset<_Elem, _Bits> _Buf;
     const typename _Istr_t::sentry _Ok(_Istr);
 
     if (_Ok) { // valid stream, extract elements
         _TRY_IO_BEGIN
-        _Str._Resize_and_overwrite(_Bits, [&, _Elem0, _Elem1](char* _Buf, size_t _Max_count) {
-            typename _Tr::int_type _Meta = _Istr.rdbuf()->sgetc();
-            for (; _Count < _Max_count; _Meta = _Istr.rdbuf()->snextc(), (void) ++_Count) {
-                // test _Meta
-                if (_Tr::eq_int_type(_Tr::eof(), _Meta)) { // end of file, quit
-                    _State |= _Istr_t::eofbit;
-                    break;
-                }
-
-                const _Elem _Char = _Tr::to_char_type(_Meta);
-
-                if (!_Tr::eq(_Char, _Elem0) && !_Tr::eq(_Char, _Elem1)) {
-                    break; // invalid element
-                }
-
-                if (_Str.max_size() <= _Str.size()) { // no room in string, give up (unlikely)
-                    _State |= _Istr_t::failbit;
-                    break;
-                }
-
-                // valid, append '0' or '1'
-                _Buf[_Count] = '0' + _Tr::eq(_Char, _Elem1);
+        _Buf._Allocate();
+        typename _Tr::int_type _Meta = _Istr.rdbuf()->sgetc();
+        for (; _Count < _Bits; _Meta = _Istr.rdbuf()->snextc(), (void) ++_Count) {
+            // test _Meta
+            if (_Tr::eq_int_type(_Tr::eof(), _Meta)) { // end of file, quit
+                _State |= _Istr_t::eofbit;
+                break;
             }
 
-            return _Count;
-        });
+            const _Elem _Char = _Tr::to_char_type(_Meta);
+
+            if (!_Tr::eq(_Char, _Elem0) && !_Tr::eq(_Char, _Elem1)) {
+                break; // invalid element
+            }
+
+            // valid, append '0' or '1'
+            _Buf._Buf[_Count] = '0' + _Tr::eq(_Char, _Elem1);
+        }
+
         _CATCH_IO_(_Istr_t, _Istr)
     }
 
@@ -652,7 +675,7 @@ basic_istream<_Elem, _Tr>& operator>>(basic_istream<_Elem, _Tr>& _Istr, bitset<_
     }
 
     _Istr.setstate(_State);
-    _Right = bitset<_Bits>(_Str); // convert string and store
+    _Right = bitset<_Bits>(_Buf._Buf, _Count); // convert string and store
     return _Istr;
 }