--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -620,14 +620,9 @@ basic_ostream<_Elem, _Tr>& operator<<(basic_ostream<_Elem, _Tr>& _Ostr, const bi
     const _Elem _Elem0       = _Ctype_fac.widen('0');
     const _Elem _Elem1       = _Ctype_fac.widen('1');
 
-    if constexpr (_Bits * sizeof(_Elem) <= _STD_BITSET_TO_STREAM_STACK_RESERVATION) {
-        _Elem _Buf[_Bits + 1];
-        _Right._To_string(_Buf, _Bits, _Elem0, _Elem1);
-        _Buf[_Bits] = _Elem{'\0'};
-        return _Ostr << _Buf;
-    } else {
-        return _Ostr << _Right.template to_string<_Elem, _Tr, allocator<_Elem>>(_Elem0, _Elem1);
-    }
+    _Buffer_for_bitset<_Elem, _Bits> _Buf;
+    _Right._To_string(_Buf, _Bits, _Elem0, _Elem1);
+    _Ostr.write(_Buf._Buf, _Bits);
 }
 
 _EXPORT_STD template <class _Elem, class _Tr, size_t _Bits>