--- a/stl/inc/bitset
+++ b/stl/inc/bitset
@@ -42,7 +42,7 @@ __declspec(noalias) bool __stdcall __std_bitset_from_string_2(void* _Dest, const
 _STD_BEGIN
 template <class _Elem, size_t _Bits, bool _Dynamic = (_Bits * sizeof(_Elem) > _STD_BITSET_TO_STREAM_STACK_RESERVATION)>
 struct _Buffer_for_bitset {
-    _Buffer_for_bitset()                                     = default;
+    _Buffer_for_bitset() noexcept                            = default;
     _Buffer_for_bitset(const _Buffer_for_bitset&)            = delete;
     _Buffer_for_bitset& operator=(const _Buffer_for_bitset&) = delete;
 
@@ -59,11 +59,11 @@ struct _Buffer_for_bitset {
 
 template <class _Elem, size_t _Bits>
 struct _Buffer_for_bitset<_Elem, _Bits, false> {
-    _Buffer_for_bitset() { /* intentionally not initializing _Buf */ }
+    _Buffer_for_bitset() noexcept { /* intentionally not initializing _Buf */ }
     _Buffer_for_bitset(const _Buffer_for_bitset&)            = delete;
     _Buffer_for_bitset& operator=(const _Buffer_for_bitset&) = delete;
 
-    void _Allocate() {}
+    void _Allocate() noexcept {}
 
     _Elem _Buf[(_Bits != 0) ? _Bits : 1];
 };