@StephanTLavavej 2024-10-17
Thanks! I pushed a conflict-free merge with `main` and several changes, including an important rename and a fix for the `_Dynamic` criterion, please double-check.

Final results on my 5950X:

Benchmark                           |   Before |    After | Speedup
:-----------------------------------|---------:|---------:|----:
`bitset_from_stream<15, char>`      | 24527 ns | 16516 ns | 1.49
`bitset_from_stream<16, char>`      | 26691 ns | 15469 ns | 1.73
`bitset_from_stream<36, char>`      | 18116 ns |  9761 ns | 1.86
`bitset_from_stream<64, char>`      | 14495 ns |  7542 ns | 1.92
`bitset_from_stream<512, char>`     |  8739 ns |  5392 ns | 1.62
`bitset_from_stream<2048, char>`    |  9585 ns |  5120 ns | 1.87
`bitset_from_stream<15, wchar_t>`   | 26066 ns | 18357 ns | 1.42
`bitset_from_stream<16, wchar_t>`   | 28599 ns | 16726 ns | 1.71
`bitset_from_stream<36, wchar_t>`   | 18729 ns | 10308 ns | 1.82
`bitset_from_stream<64, wchar_t>`   | 14788 ns |  8018 ns | 1.84
`bitset_from_stream<512, wchar_t>`  |  9561 ns |  5909 ns | 1.62
`bitset_from_stream<2048, wchar_t>` |  9118 ns |  5181 ns | 1.76