@AlexGuteniev 2024-10-18
> an important rename and a fix for the `_Dynamic` criterion

I don't agree. I wanted the constant stack size (like `basic_string` SSO, and like it was for `to_string` before PR), and I don't see how array size multiplied by element size wouldn't get that.