@StephanTLavavej 2024-10-18
Ah - yes, at a minimum, it would make sense to not consume more stack space than necessary. (I think I missed that at first because the parameter wasn't named `_Bits`, and I didn't fully rethink the meaning after doing the rename).

If you additionally want to adjust the criterion so that we switch to dynamic once the size would exceed a number of bytes instead of a number of elements, go ahead - I'll move this back to WIP. Thanks!