@AlexGuteniev 2024-10-18
Applied changes, verified that benchmark results are about the same (although there is seemingly random variation)