@StephanTLavavej 2024-10-21
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.