@StephanTLavavej 2024-10-21
Thanks for optimizing this operator! :rocket: :stopwatch: :tada: